+++ org/postgresql/jdbc4/Jdbc4Blob.java
@@ -3,7 +3,7 @@
 * Copyright (c) 2004-2005, PostgreSQL Global Development Group
 *
 * IDENTIFICATION
-*   $PostgreSQL$
+*   $PostgreSQL: pgjdbc/org/postgresql/jdbc4/Jdbc4Blob.java,v 1.1 2006/06/08 10:34:52 jurka Exp $
 *
 *-------------------------------------------------------------------------
 */
@@ -15,7 +15,7 @@ import java.sql.*;
 public class Jdbc4Blob extends AbstractJdbc4Blob implements java.sql.Blob
 {
 
-    public Jdbc4Blob(org.postgresql.PGConnection conn, int oid) throws SQLException
+    public Jdbc4Blob(org.postgresql.PGConnection conn, long oid) throws SQLException
     {
         super(conn, oid);
     }
